\name{anova.RidgeBinaryLogistic}
\alias{anova.RidgeBinaryLogistic}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compares two binary logistic models
}
\description{
Anova for comparing two binary logistic models
}
\usage{
\method{anova}{RidgeBinaryLogistic}(object, object2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The first model
}
  \item{object2}{
The second model
}
  \item{...}{
Any additional arguments
}
}
\details{
Anova for comparing two binary logistic models
}
\value{
The comparison of the two models.
}

\author{
Jose Luis Vicente Villardon
}

\examples{
# Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
