\name{Biplot.PLSR1BIN}
\alias{Biplot.PLSR1BIN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Biplot for a PLSR model with a binary response
}
\description{
Biplot for a PLSR model with a binary response
}
\usage{
Biplot.PLSR1BIN(plsr)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{plsr}{
An object of class PLSR1BIN.
}
}
\details{
Biplot for a PLSR model with a binary response
} 
\value{
The biplot for the independent variables with the response as supplementary binary variable.
}
\references{
Ugarte-Fajardo, J., Bayona-Andrade, O., Criollo-Bonilla, R., Cevallos-Cevallos, J., Mariduena-Zavala, M., Ochoa-Donoso, D., & Vicente-Villardon, J. L. (2020). Early detection of black Sigatoka in banana leaves using hyperspectral images. Applications in plant sciences, 8(8), e11383.
}
\author{
Jose Luis Vicente Villardon
}

\seealso{
\code{\link{PLSR1Bin}}
}
\examples{
# Not Yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
