\name{Chemical}
\alias{Chemical}
\docType{data}
\title{
Chemical data 
}
\description{
Ecological data
}
\usage{data("Chemical")}
\format{
  A data frame with 324 observations on the following 16 variables.
  \describe{
    \item{\code{Treatment}}{a factor with levels \code{F0N0} \code{F0N1} \code{F0N2} \code{F0N3} \code{F1N0} \code{F1N1} \code{F1N2} \code{F1N3} \code{F2N0} \code{F2N1} \code{F2N2} \code{F2N3}}
    \item{\code{FISH}}{a factor with levels \code{F0} \code{F1} \code{F2}}
    \item{\code{NUTRIENTS}}{a factor with levels \code{N0} \code{N1} \code{N2} \code{N3}}
    \item{\code{WEEKS}}{a factor with levels \code{W1} \code{W2} \code{W3} \code{W4} \code{W5} \code{W6} \code{W7} \code{W8} \code{W9}}
    \item{\code{TEMPERATURE}}{a numeric vector}
    \item{\code{pH}}{a numeric vector}
    \item{\code{ALKALINITYmeql}}{a numeric vector}
    \item{\code{CO2free}}{a numeric vector}
    \item{\code{NNH4mgl}}{a numeric vector}
    \item{\code{NNO3mgl}}{a numeric vector}
    \item{\code{SRPmglP}}{a numeric vector}
    \item{\code{TPmgl}}{a numeric vector}
    \item{\code{TSSmgl}}{a numeric vector}
    \item{\code{CONDUCTIVITYmScm}}{a numeric vector}
    \item{\code{TSPmglP}}{a numeric vector}
    \item{\code{Chlorophyllamgl}}{a numeric vector}
  }
}
\details{
Chemical Data
}
\source{
Department of Ecology. University of Leon. (Spain)
}
\references{
To add
}
\examples{
data(Chemical)
## maybe str(Chemical) ; plot(Chemical) ...
}
\keyword{datasets}
