\name{ConfidenceInterval}
\alias{ConfidenceInterval}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Confidence Interval for the mean
}
\description{
Calculates Confidence Interval for the mean of a Numerical Variable.
}
\usage{
ConfidenceInterval(x, Desv = NULL, df = NULL, Confidence = 0.95)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The numerical variable
}
  \item{Desv}{
Standard deviation. If NULL, the sd is calculated from the data
}
  \item{df}{
Degrees of freedom
}
  \item{Confidence}{
Confidence Level
}
}
\details{
Calculates Confidence Interval for the mean of a Numerical Variable. 
}
\value{
The confidence Interval for the mean
}
\author{
Jose Luis Vicente Villardon
}

\examples{
# Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
