\name{Convert3wArray2List}
\alias{Convert3wArray2List}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Converts a three way array into a list
}
\description{
Converts a three way array into a list
}
\usage{
Convert3wArray2List(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
A three way array
}
}
\details{
Converts a three way array into a list
}
\value{
A list
}

\author{
Jose Luis Vicente-Villardon
}

\examples{
#No examples yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
