\name{ConvertFactors2Integers}
\alias{ConvertFactors2Integers}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert a factor to integer numbers
}
\description{
Convert a factor to integer numbers
}
\usage{
ConvertFactors2Integers(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector with a factor
}
}
\details{
Convert a factor to integer numbers
}
\value{
a vector with the converted values
}

\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----
}
