\name{ConvertList23wArray}
\alias{ConvertList23wArray}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Converts a list of matrices into a three way array
}
\description{
Converts a list of matrices into a three way array. All the matrices in the list must have the same size.
}
\usage{
ConvertList23wArray(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
 A list with data matrices.
}
}
\details{
Converts a list of matrices into a three way array. All the matrices in the list must have the same size.
}
\value{
A three-way array
}

\author{
Jose Luis Vicente-Villardon
}

\examples{
# No examples yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
