\name{moth}
\alias{moth}
\docType{data}
\title{
Moth data
}
\description{
Moth data
}
\usage{data("moth")}
\format{
  A data frame with 12 observations on the following 14 variables.
  \describe{
    \item{\code{s1}}{a numeric vector}
    \item{\code{s2}}{a numeric vector}
    \item{\code{s3}}{a numeric vector}
    \item{\code{s4}}{a numeric vector}
    \item{\code{s5}}{a numeric vector}
    \item{\code{s6}}{a numeric vector}
    \item{\code{s7}}{a numeric vector}
    \item{\code{s8}}{a numeric vector}
    \item{\code{s9}}{a numeric vector}
    \item{\code{s10}}{a numeric vector}
    \item{\code{s11}}{a numeric vector}
    \item{\code{s12}}{a numeric vector}
    \item{\code{s13}}{a numeric vector}
    \item{\code{s14}}{a numeric vector}
  }
}
\details{
Moth data
}
\source{
Withaker
}
\references{
Application of the Parametric Bootstrap to Models that Incorporate a Singular
Value Decomposition
Luis Milan; Joe Whittaker
Applied Statistics, Vol. 44, No. 1. (1995), pp. 31-49.
}
\examples{
data(moth)
## maybe str(moth) ; plot(moth) ...
}
\keyword{datasets}
