\name{Multiquad}
\alias{Multiquad}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Multidimensional Gauss-Hermite quadrature
}
\description{
Multidimensional Gauss-Hermite quadrature
}
\usage{
Multiquad(nnodes, dims)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nnodes}{
Number of nodes of the quadrature
}
  \item{dims}{
Dimension of the solution
}
}
\details{
Multidimensional Gauss-Hermite quadrature
}
\value{
Multidimensional Gauss-Hermite quadrature
}
\references{
Jackel, P. (2005). A note on multivariate Gauss-Hermite quadrature. http://www.awdz65.dsl.pipex.com/ANoteOnMultivariateGaussHermiteQuadrature.pdf
}
\author{
Jose Luis Vicente Villardon
}

\examples{
Multiquad(5, 3)
}
