\name{MultiTableStatistics}
\alias{MultiTableStatistics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Statistics for multiple tables
}
\description{
Statistics for multiple tables
}
\usage{
MultiTableStatistics(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
A multiple table
}
}
\details{
Statistics for multiple tables
}
\value{
A list with vectors of statistics for each table
}

\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----
}
