\name{MultiTableTransform}
\alias{MultiTableTransform}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Initial Transformation of a multi table object
}
\description{
Initial Transformation of a multi table object
}
\usage{
MultiTableTransform(X, InitTransform = "Standardize columns")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
Multi-table object
}
  \item{InitTransform}{
Initial Transformattion
}
}
\details{
Initial Transformation of a multi table object
}
\value{
he table transformed
}
\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----

}
