\name{ones}
\alias{ones}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Matrix of ones
}
\description{
Square matrix of ones
}
\usage{
ones(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
Order of the matrix
}
}
\details{
Square matrix of ones
}
\value{
A matrix of ones of order n.
}

\author{
Jose Luis Vicente Villardon
}

\examples{
ones(6)
}

