\name{OrthogonalizeScores}
\alias{OrthogonalizeScores}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Orthogonalize a set of Scores calculated by other procedure
}
\description{
Orthogonalize a set of Scores calculated by other procedure
}
\usage{
OrthogonalizeScores(scores)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{scores}{
A matrix containing the scores
}
}
\details{
Orthogonalize a set of Scores calculated by other procedure proyecting onto the dimensions defined by the eigenvectors of the covariance matrix
}
\value{
The orthogonalised scores.
}

\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----

}