\name{plot.CVA}
\alias{plot.CVA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot of a Canonical Variate Analysis
}
\description{
Plot of a Canonical Variate Analysis
}
\usage{
\method{plot}{CVA}(x, A1 = 1, A2 = 2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Object of class CVA
}
  \item{A1}{
Dimension for the first axis of the representation
}
  \item{A2}{
Dimension for the second axis of the representation
}
  \item{\dots}{
Additional arguments
}
}
\details{
Plot of a Canonical Variate Analysis
}
\value{
Te Vanonical variate plot
}
\author{
Jose Luis Vicente Villardon
}
