\name{PoliticalFigures}
\alias{PoliticalFigures}
\docType{data}
\title{
Political Figures in the USA
}
\description{
Does the American public actively differentiate political stimuli along ideological lines?. Dissimilarities among 13 political figurein the USA.
}
\usage{data("PoliticalFigures")}
\format{
  A data frame with the dissimilarities among 13 political figures in the USA.
  \describe{
    \item{\code{G._W._Bush}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{John_Kerry}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Ralph_Nader}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Dick_Cheney}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{John_Edwards}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Laura_Bush}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Hillary_Clinton}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Bill_Clinton}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Colin_Powell}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{John_Ashcroft}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{John_McCain}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Democ._Party}}{a numeric vector with the dissimilarities with the other figures}
    \item{\code{Repub._Party}}{a numeric vector with the dissimilarities with the other figures}
  }
}
\details{
We have taken information from the 2004 CPS American National Election Study. 
Specifically 711 NES respondents' feeling thermometer ratings of thirteen prominent political
figures from the period of the 2004 election: George W. Bush; John Kerry; Ralph Nader;
Richard Cheney; John Edwards; Laura Bush; Hillary Clinton; Bill Clinton; Colin Powell;
John Ashcroft; John McCain; the Democratic party; and the Republican party. 
With the respondent scores, a dissimilarity among each pair of figures 
}
\source{
Jacoby, W. G., & Armstrong, D. A. (2014). Bootstrap Confidence Regions for Multidimensional Scaling Solutions. American Journal of Political Science, 58(1), 264-278.
}
\references{
Jacoby, W. G., & Armstrong, D. A. (2014). Bootstrap Confidence Regions for Multidimensional Scaling Solutions. American Journal of Political Science, 58(1), 264-278.
}
\examples{
# Not yet
}
\keyword{datasets}
