\name{print.MGC}
\alias{print.MGC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prints the results of Model-Based Gaussian Clustering algorithms
}
\description{
Prints the results of Model-Based Gaussian Clustering algorithms
}
\usage{
\method{print}{MGC}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An object of class "MGC"
}
  \item{\dots}{
Any aditional parameters
}
}
\details{
Prints the results of Model-Based Gaussian Clustering algorithms
}
\value{
No value returned
}

\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
