\name{print.RidgeBinaryLogistic}
\alias{print.RidgeBinaryLogistic}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
prints an object of class RidgeBinaryLogistic
}
\description{
prints an object of class RidgeBinaryLogistic
}
\usage{
\method{print}{RidgeBinaryLogistic}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An object of class 
}
  \item{\dots}{
Aditional Arguments
}
}
\details{
Prints an object of class RidgeBinaryLogistic
}
\value{
The main resuls of a binary logistic regression
}

\author{
Jose Luis Vicente Villardon
}

\examples{
# Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
