\name{RAPD}
\alias{RAPD}
\docType{data}
\title{
Sugar Cane Data
}
\description{
Molecular characteristics of 50 varieties of sugar cane.
}
\usage{data(RAPD)}
\format{
  A data frame with 50 observations on 168 variables. 1-120: Random aplified polymorphic DNA and 121-168: Microsatellites

}
\details{
Dta are codified as presence or absence of the dominant marker.
}

\examples{
data(RAPD)
## maybe str(RAPD) ; plot(RAPD) ...
}
\keyword{datasets}
