\name{RemoveRowsWithNaNs}
\alias{RemoveRowsWithNaNs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Remove rows that contains NaNs (missing data)
}
\description{
Remove rows that contains NaNs to obtain a matrix wothout missind data
}
\usage{
RemoveRowsWithNaNs(x, cols = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The matrix to be arranged
}
  \item{cols}{
A set of columns to check as a vector of integers
}
}
\details{
Remove rows that contains NaNs to obtain a matrix wothout missind data
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{x}{Matrix without missing data}

}

\author{
Jose Luis Vicente-Villardon
}

