\name{riano}
\alias{riano}
\docType{data}
\title{
Ecological data from Riano (Spain)
}
\description{
Ecological data from Riano (Spain)
}
\usage{data("riano")}
\format{
  A data frame with 70 observations on the following 25 variables.
  \describe{
    \item{\code{Week}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J}}
    \item{\code{Depth}}{a factor with levels \code{0} \code{2} \code{5} \code{10} \code{15} \code{20} \code{Bottom}}
    \item{\code{Cianof}}{a numeric vector}
    \item{\code{Crisof}}{a numeric vector}
    \item{\code{Haptof}}{a numeric vector}
    \item{\code{Crasp}}{a numeric vector}
    \item{\code{Cripto}}{a numeric vector}
    \item{\code{Dinof}}{a numeric vector}
    \item{\code{Diatom}}{a numeric vector}
    \item{\code{Euglen}}{a numeric vector}
    \item{\code{Prasin}}{a numeric vector}
    \item{\code{Clorof}}{a numeric vector}
    \item{\code{Zigofi}}{a numeric vector}
    \item{\code{Xantof}}{a numeric vector}
    \item{\code{malgas}}{a numeric vector}
    \item{\code{Ta}}{a numeric vector}
    \item{\code{X02}}{a numeric vector}
    \item{\code{pH}}{a numeric vector}
    \item{\code{COND}}{a numeric vector}
    \item{\code{SiO2}}{a numeric vector}
    \item{\code{P.PO4}}{a numeric vector}
    \item{\code{Chla}}{a numeric vector}
    \item{\code{Chlb}}{a numeric vector}
    \item{\code{Chlc}}{a numeric vector}
    \item{\code{IM}}{a numeric vector}
  }
}
\details{
Ecological data from Riano (Spain). Abundance of several algae taxonomic groups and several environmental variables
}
\source{
Department of Ecology. University of Leon. Spain
}

\examples{
data(riano)
## maybe str(riano) ; plot(riano) ...
}
\keyword{datasets}
