\name{smoking}
\alias{smoking}
\docType{data}
\title{
Smoking habits
}
\description{
Frequency table representing smoking habits of different employees in a company 
}
\usage{data(smoking)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{None}}{a numeric vector}
    \item{\code{Light}}{a numeric vector}
    \item{\code{Medium}}{a numeric vector}
    \item{\code{Heavy}}{a numeric vector}
  }
}
\details{
Frequency table representing smoking habits of different employees in a company 
}
\source{
http://orange.biolab.si/docs/latest/reference/rst/Orange.projection.correspondence/
}
\references{
Greenacre, Michael (1983). Theory and Applications of Correspondence Analysis. London: Academic Press.
}
\examples{
data(smoking)
## maybe str(smoking) ; plot(smoking) ...
}
\keyword{datasets}
