\name{spiders}
\alias{spiders}
\docType{data}
\title{
Hunting Spiders Data 
}
\description{
Hunting spiders data transformed into Presence/Abscense.
}
\usage{data(spiders)}
\format{
 A data frame with 28 observations of presence/absence of 12 hunting spider species
  \describe{
    \item{Alopacce}{Presence/Absence of the species Alopecosa accentuata}
    \item{Alopcune}{Presence/Absence of the species Alopecosa cuneata}
    \item{Alopfabr}{Presence/Absence of the species Alopecosa fabrilis}
    \item{Arctlute}{Presence/Absence of the species Arctosa lutetiana}
    \item{Arctperi}{Presence/Absence of the species Arctosa perita}
    \item{Auloalbi}{Presence/Absence of the species Aulonia albimana}
    \item{Pardlugu}{Presence/Absence of the species Pardosa lugubris}
    \item{Pardmont}{Presence/Absence of the species Pardosa monticola}
    \item{Pardnigr}{Presence/Absence of the species Pardosa nigriceps}
    \item{Pardpull}{Presence/Absence of the species Pardosa pullata}
    \item{Trocterr}{Presence/Absence of the species Trochosa terricola}
    \item{Zoraspin}{Presence/Absence of the species Zora spinimana}
  }
}
\source{
van der Aart, P. J. M., and Smeenk-Enserink, N. (1975) Correlations between distributions of hunting spiders (Lycos- idae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology 25, 1-45.
}
\examples{
data(spiders)
}
\keyword{datasets}

