\name{SpidersEnv}
\alias{SpidersEnv}
\docType{data}
\title{
Hunting spiders environmental data.
}
\description{
Hunting spiders environmental data.
}
\usage{data("SpidersEnv")}
\format{
  A data frame with 28 observations on the following 6 variables.
  \describe{
    \item{\code{Watcont}}{Water content}
    \item{\code{Barsand}}{Bare sand}
    \item{\code{Covmoss}}{Cover moss}
    \item{\code{Ligrefl}}{Light reflection}
    \item{\code{Falltwi}}{Fallen Twings}
    \item{\code{Coverher}}{Cover Herbs}
  }
}
\details{
Hunting spiders environmental data.
}
\source{
van der Aart, P. J. M., and Smeenk-Enserink, N. (1975) Correlations between distributions of hunting spiders (Lycos- idae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology 25, 1-45.
}
\references{
Ter Braak, C. J. (1986). Canonical correspondence analysis: a new eigenvector technique for multivariate direct gradient analysis. Ecology, 67(5), 1167-1179.
}
\examples{
data(SpidersEnv)
## maybe str(SpidersEnv) ; plot(SpidersEnv) ...
}
\keyword{datasets}
