\name{SpidersSp}
\alias{SpidersSp}
\docType{data}
\title{
Hunting Spiders Data 
}
\description{
Hunting spiders abundances data.
}
\usage{data("SpidersSp")}
\format{
 A data frame with 28 observations of abundance of 12 hunting spider species
  \describe{
    \item{Alopacce}{Abundance of the species Alopecosa accentuata}
    \item{Alopcune}{Abundance of the species Alopecosa cuneata}
    \item{Alopfabr}{Abundance of the species Alopecosa fabrilis}
    \item{Arctlute}{Abundance of the species Arctosa lutetiana}
    \item{Arctperi}{Abundance of the species Arctosa perita}
    \item{Auloalbi}{Abundance of the species Aulonia albimana}
    \item{Pardlugu}{Abundance of the species Pardosa lugubris}
    \item{Pardmont}{Abundance of the species Pardosa monticola}
    \item{Pardnigr}{Abundance of the species Pardosa nigriceps}
    \item{Pardpull}{Abundance of the species Pardosa pullata}
    \item{Trocterr}{Abundance of the species Trochosa terricola}
    \item{Zoraspin}{Abundance of the species Zora spinimana}
  }
}

\source{
van der Aart, P. J. M., and Smeenk-Enserink, N. (1975) Correlations between distributions of hunting spiders (Lycos- idae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology 25, 1-45.
}
\references{
Ter Braak, C. J. (1986). Canonical correspondence analysis: a new eigenvector technique for multivariate direct gradient analysis. Ecology, 67(5), 1167-1179.
}
\examples{
data(SpidersSp)
## maybe str(SpidersSp) ; plot(SpidersSp) ...
}
\keyword{datasets}
