\name{SSI3w}
\alias{SSI3w}
\docType{data}
\title{
Sustainability Society Index (3w)
}
\description{
Sustainability Society Index, Three way table
}
\usage{data("SSI3w")}
\format{
  The format is:
List of 6
 $ a2006: num [1:154, 1:21] 10 9.3 6.6 10 8.9 10 10 10 8.3 10 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:21] "Sufficient_Food" "Sufficient_to_Drink" "Safe_Sanitation" "Education_" ...
 $ a2008: num [1:154, 1:21] 10 9.4 7.1 10 9.3 10 10 10 8.3 10 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:21] "Sufficient_Food" "Sufficient_to_Drink" "Safe_Sanitation" "Education_" ...
 $ a2010: num [1:154, 1:21] 10 9.4 7.7 10 9.4 10 10 10 8.3 10 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:21] "Sufficient_Food" "Sufficient_to_Drink" "Safe_Sanitation" "Education_" ...
 $ a2012: num [1:154, 1:21] 10 10 8.1 10 9.3 10 10 10 8.3 10 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:21] "Sufficient_Food" "Sufficient_to_Drink" "Safe_Sanitation" "Education_" ...
 $ a2014: num [1:154, 1:21] 10 10 8.4 10 9.3 10 10 10 8.3 10 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:21] "Sufficient_Food" "Sufficient_to_Drink" "Safe_Sanitation" "Education_" ...
 $ a2016: num [1:154, 1:21] 10 10 8.6 10 9.4 10 10 10 8.4 10 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:21] "Sufficient_Food" "Sufficient_to_Drink" "Safe_Sanitation" "Education_" ...
}
\details{
Sustainability Society Index
}
\source{
https://ssi.wi.th-koeln.de
}
\references{
Gallego-Alvarez, I., Galindo-Villardon, M. P., & Rodriguez-Rosa, M. (2015). Analysis of the Sustainable Society Index Worldwide: A Study from the Biplot Perspective. Social Indicators Research, 120(1), 29-65. https://doi.org/10.1007/s11205-014-0579-9
}
\examples{
data(SSI3w)
## maybe str(SSI3w) ; plot(SSI3w) ...
}
\keyword{datasets}
