\name{SSIEcon3w}
\alias{SSIEcon3w}
\docType{data}
\title{
Sustainability Society Index
}
\description{
Sustainability Society Index
}
\usage{data("SSIEcon3w")}
\format{
  The format is:
List of 6
 $ a2006: num [1:154, 1:5] 1.2 1 1 4.6 1 5.4 9.9 1.9 1 1 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:5] "Organic_Farming" "Genuine_Savings" "GDP" "Employment" ...
 $ a2008: num [1:154, 1:5] 1 1 1 4.2 1 5.6 9.9 1.9 1 1 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:5] "Organic_Farming" "Genuine_Savings" "GDP" "Employment" ...
 $ a2010: num [1:154, 1:5] 1.1 1 1 5.8 1.1 5.6 9.9 2 1 1 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:5] "Organic_Farming" "Genuine_Savings" "GDP" "Employment" ...
 $ a2012: num [1:154, 1:5] 1.1 1 1 5.7 1.1 5.7 9.9 2 1 1 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:5] "Organic_Farming" "Genuine_Savings" "GDP" "Employment" ...
 $ a2014: num [1:154, 1:5] 1.1 1 1 5.3 1.1 5.7 9.9 2.1 1.2 1 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:5] "Organic_Farming" "Genuine_Savings" "GDP" "Employment" ...
 $ a2016: num [1:154, 1:5] 1.1 1 1 4.8 1.1 6.8 9.9 2 1.2 1 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:5] "Organic_Farming" "Genuine_Savings" "GDP" "Employment" ...
}
\details{
Sustainability Society Index
}
\source{
https://ssi.wi.th-koeln.de
}
\references{
Gallego-Alvarez, I., Galindo-Villardon, M. P., & Rodriguez-Rosa, M. (2015). Analysis of the Sustainable Society Index Worldwide: A Study from the Biplot Perspective. Social Indicators Research, 120(1), 29-65. https://doi.org/10.1007/s11205-014-0579-9
}
\examples{
data(SSIEcon3w)
## maybe str(SSIEcon3w) ; plot(SSIEcon3w) ...
}
\keyword{datasets}
