\name{SSIEnvir3w}
\alias{SSIEnvir3w}
\docType{data}
\title{
Sustainability Society Index
}
\description{
Sustainability Society Index
}
\usage{data("SSIEnvir3w")}
\format{
  The format is:
List of 6
 $ a2006: num [1:154, 1:7] 4.2 6.5 4 4.9 7.7 5.7 8.1 4.9 2.8 6.3 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:7] "Biodiversity_" "Renewable_Water_Resources" "Consumption" "Energy_Use" ...
 $ a2008: num [1:154, 1:7] 4.8 6.5 4 5.1 7.7 5.7 8 5.7 2.8 6 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:7] "Biodiversity_" "Renewable_Water_Resources" "Consumption" "Energy_Use" ...
 $ a2010: num [1:154, 1:7] 5.4 6.6 4 5.2 7.7 5.7 8 6.4 2.8 5.8 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:7] "Biodiversity_" "Renewable_Water_Resources" "Consumption" "Energy_Use" ...
 $ a2012: num [1:154, 1:7] 5.3 6.6 4 5.3 7.7 6.1 8 6.8 2.8 5.8 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:7] "Biodiversity_" "Renewable_Water_Resources" "Consumption" "Energy_Use" ...
 $ a2014: num [1:154, 1:7] 5.6 6.6 4 5.3 7.7 7 7.9 7.3 2.8 6 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:7] "Biodiversity_" "Renewable_Water_Resources" "Consumption" "Energy_Use" ...
 $ a2016: num [1:154, 1:7] 5.5 6.6 4.1 5.4 7.8 7.3 7.9 7.3 2.9 5.9 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:154] "Albania" "Algeria" "Angola" "Argentina" ...
  .. ..$ : chr [1:7] "Biodiversity_" "Renewable_Water_Resources" "Consumption" "Energy_Use" ...
}
\details{
Sustainability Society Index
}
\source{
https://ssi.wi.th-koeln.de
}
\references{
Gallego-Alvarez, I., Galindo-Villardon, M. P., & Rodriguez-Rosa, M. (2015). Analysis of the Sustainable Society Index Worldwide: A Study from the Biplot Perspective. Social Indicators Research, 120(1), 29-65. https://doi.org/10.1007/s11205-014-0579-9
}
\examples{
data(SSIEnvir3w)
## maybe str(SSIEnvir3w) ; plot(SSIEnvir3w) ...
}
\keyword{datasets}
