\name{summary.CCA.sol}
\alias{summary.CCA.sol}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of the solution of a CCA 
}
\description{
Summary of the solution of a CCA 
}
\usage{
\method{summary}{CCA.sol}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class CCA.sol
}
  \item{...}{
Aditional arguments
}
}
\details{
Summary of the solution of a CCA 
}
\value{
The main results of a CCA
}

\author{
Jose Luis Vicente Villardon
}

\seealso{
\code{\link{CCA}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
