\name{summary.CVA}
\alias{summary.CVA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of a Canonical Variate Analysis
}
\description{
Summary of a Canonical Variate Analysis
}
\usage{
\method{summary}{CVA}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class CVA
}
  \item{\dots}{
Any aditional arguments
}
}
\details{
Summary of a Canonical Variate Analysis
}
\value{
The summary
}

\author{
Jose Luis Vicente Villardon
}
\examples{
# Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
