\name{summary.MGC}
\alias{summary.MGC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of Model-Based Gaussian Clustering results
}
\description{
Summarizes the results of Model-Based Gaussian Clustering algorithms
}
\usage{
\method{summary}{MGC}(object, Centers = TRUE, Covariances = TRUE, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class "MGC"
}
  \item{Centers}{
Should the Centers be shown
}
  \item{Covariances}{
Should the Covariances be shown
}
  \item{\dots}{
Any aditional Parameters
}
}
\details{
Summarizes the results of Model-Based Gaussian Clustering algorithms
}
\value{
No value returned
}

\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as

}
