\name{summary.PCA.Analysis}
\alias{summary.PCA.Analysis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of the results of a PCA.
}
\description{
Sumarizes the results of a PCA Analysis.
}
\usage{
\method{summary}{PCA.Analysis}(object, latex = FALSE, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The object with the results of s PCA Analysis.
}
  \item{latex}{
Should return latex tables?
}
  \item{\dots}{
Aditional arguments.
}
}
\details{
Sumarizes the results of a PCA Analysis, including latex tables for presentation.
}
\value{
A summary of the main results
}

\author{
Jose Luis Vicente Villardon
}

\examples{
# Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
