\name{summary.PCA.Bootstrap}
\alias{summary.PCA.Bootstrap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of a PCA.Bootstrap object
}
\description{
Summary of a PCA.Bootstrap object
}
\usage{
\method{summary}{PCA.Bootstrap}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class PCA.Bootstrap 
}
  \item{\dots}{
Additional arguments
}
}
\details{
Summary of a PCA.Bootstrap object
}
\value{
The summary
}
\author{
Jose Luis Vicente Villardon
}

