\name{summary.PLSR}
\alias{summary.PLSR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of a PLSR object
}
\description{
Summary of a PLSR object
}
\usage{
\method{summary}{PLSR}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class PLSR
}
  \item{\dots}{
Additional arguments
}
}
\details{
Summary of a PLSR object
}
\value{
The summary of the object
}

\author{
Jose Luis Vicente Villardon
}
