\name{summary.PLSR1Bin}
\alias{summary.PLSR1Bin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of PLSR with a Binary Response
}
\description{
Summary of PLSR with a single binary Response
}
\usage{
\method{summary}{PLSR1Bin}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class PLSR1Bin
}
  \item{\dots}{
Aditional arguments
}
}
\details{
Summary of PLSR with a single binary Response
}
\value{
The summary
}

\author{
Jose Luis Viecente Villlardon
}

\examples{
#Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
