\name{summary.RidgeBinaryLogistic}
\alias{summary.RidgeBinaryLogistic}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of a Binary Logistic Regression with Ridge Penalization
}
\description{
Summarizes the results of a Binary Logistic Regression with Ridge Penalization
}
\usage{
\method{summary}{RidgeBinaryLogistic}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The object with te results of the logistic regression.
}
  \item{\dots}{
Any other arguments
}
}
\details{
Summarizes the results of a Binary Logistic Regression with Ridge Penalization.
}
\value{
The summary
}

\author{
Jose Luis Vicente Villardon
}

\examples{
# Not Yet
}
