\name{summary.TetraDualStatis}
\alias{summary.TetraDualStatis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary of the results of TetraDualStatis
}
\description{
Summary of the results of TetraDualStatis
}
\usage{
\method{summary}{TetraDualStatis}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The result of a Tetra Dual Statis Analysis
}
  \item{\dots}{
aditional arguments
}
}
\details{
Summarizes the results of TetradUalStatis
}
\value{
No value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
Laura Vicente-Gonzalez, José Luis Vicente-Villardon
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
# No examples yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
