\name{ThreeWay2FrontalSlices}
\alias{ThreeWay2FrontalSlices}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Three to two way data
}
\description{
Three to two way data.
}
\usage{
ThreeWay2FrontalSlices(X, Slice = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
A three way array.
}
  \item{Slice}{
The mode for the rows
}
}
\details{
Three to two way data. The provided mode is placen on the rows. The columns are the result of intercatively coding the other two modes.
}
\value{
A two way matrix.
}

\author{
José Luis Vicente- Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
