\name{wa}
\alias{wa}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
Extracts the weighted averages of a CCA solution
}
\description{
Extracts the weighted averages of a CCA solution
}
\usage{
wa(CCA.sol, transformed = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{CCA.sol}{
The solution of a CCA
}
  \item{transformed}{
Average of the transformed or the original data?
}
}
\details{
Extracts the weighted averages of a CCA solution
}
\value{
A matrix with the averages
}
\author{
icente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----
}
