\name{wcor}
\alias{wcor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Weighted correlations
}
\description{
Weighted correlations
}
\usage{
wcor(d1, d2, w = rep(1, nrow(d1))/nrow(d1))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{d1}{
First Vector
}
  \item{d2}{
Second vector to correlate
}
  \item{w}{
weights for ecah element of the vectors
}
}
\details{
Weighted correlations
}
\value{
Weighted correlation
}

\author{
Jose Luis Vicente Villardon
}
