\name{weighted.quantile}
\alias{weighted.quantile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Weighted quantiles
}
\description{
Weighted quantiles
}
\usage{
weighted.quantile(x, w, q = 0.5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The numerical variable.
}
  \item{w}{
Weights
}
  \item{q}{
Quantile
}
}

\value{
The quantile
}

\author{
Jose Luis Vicente Villardon
}

\examples{
##---- Should be DIRECTLY executable !! ----

}
