\name{zeros}
\alias{zeros}

\title{
Matrix of zeros as in Matlab
}
\description{
Matrix of zeros
}
\usage{
zeros(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
Dimension of the matrix
}
}

\value{
A matrix of zeros
}

\author{
Jose Luis Vicente Villardon
}

\examples{
zeros(6)
}
