\name{logit}
\alias{logit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Logit function
}
\description{
Calculates the logit of a probability
}
\usage{
logit(p)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{
A probability
}
}
\details{
Calculates the logit of a probability
}
\value{
The lo git of the provided probebility
}

\author{
Jose Luis Vicente Villardón
}
