\name{MultiTableTransform}
\alias{MultiTableTransform}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Initial Transformation of a multi table object
}
\description{
Initial Transformation of a multi table object
}
\usage{
MultiTableTransform(X, InitTransform = "Standardize columns", dual = FALSE,
CommonSD = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
Multi-table object
}
  \item{InitTransform}{
Initial Transformattion
}
  \item{dual}{
Is the transformation for the dual versions?
}
  \item{CommonSD}{
Should a common standard deviation be used for all the groups?
}
}
\details{
Initial Transformation of a multi table object
}
\value{
he table transformed
}
\author{
Jose Luis Vicente Villardon
}


